HNY: store the sound labels within each category
The top-level map contains categories such as "guitar" or
"silence".  Now each category is also a map with a label
for each sound, typically with note names such as "g1" or
"c3". The behaviour hasn't changed yet, but this allows
selecting notes for their musical notes instead of just
plain categories.