update vibes sounds (now tuned with guitar & voice)
The guitar and voice sounds were in tune with an old
piano which was not at 440Hz.  The vibes sounds are
from a standard concert picth 440Hz instrument, so
now they've been lowered to be in tune with the rest.
It could all be recorded again at 440Hz since the
piano with the lower pitch is not actually being used
in the end...